extract set clause visitor
